+++ src/gamspy/_symbols/universe_alias.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import os
+import threading
 from typing import TYPE_CHECKING
 
 import gams.transfer as gt
@@ -8,6 +10,7 @@ from gams.core.gdx import GMS_DT_ALIAS
 import gamspy as gp
 import gamspy._algebra.condition as condition
 import gamspy._validation as validation
+from gamspy.exceptions import ValidationError
 
 if TYPE_CHECKING:
     from gamspy import Container
@@ -66,8 +69,10 @@ class UniverseAlias(gt.UniverseAlias):
 
         return obj
 
-    def __new__(cls, container: Container, name: str = "universe"):
-        if not isinstance(container, gp.Container):
+    def __new__(
+        cls, container: Container | None = None, name: str = "universe"
+    ):
+        if container is not None and not isinstance(container, gp.Container):
             raise TypeError(
                 "Container must of type `Container` but found"
                 f" {type(container)}"
@@ -77,6 +82,11 @@ class UniverseAlias(gt.UniverseAlias):
             raise TypeError(f"Name must of type `str` but found {type(name)}")
 
         try:
+            if not container:
+                container = gp._ctx_managers[
+                    (os.getpid(), threading.get_native_id())
+                ]
+
             symbol = container[name]
             if isinstance(symbol, cls):
                 return symbol
@@ -88,9 +98,20 @@ class UniverseAlias(gt.UniverseAlias):
         except KeyError:
             return object.__new__(cls)
 
-    def __init__(self, container: Container, name: str = "universe"):
+    def __init__(
+        self, container: Container | None = None, name: str = "universe"
+    ):
         # check if the name is a reserved word
         name = validation.validate_name(name)
+        if container is None:
+            try:
+                container = gp._ctx_managers[
+                    (os.getpid(), threading.get_native_id())
+                ]
+            except KeyError as e:
+                raise ValidationError(
+                    "UniverseAlias requires a container."
+                ) from e
 
         super().__init__(container, name)