+++ src/gamspy/formulations/shape.py
@@ -37,7 +37,9 @@ def _flatten_dims_par(
     new_domain, _ = _get_new_domain(x, dims)
 
     new_shape = [len(d) for d in new_domain]
-    data = data.reshape(new_shape)
+
+    if data is not None:
+        data = data.reshape(new_shape)
 
     out = m.addParameter(domain=new_domain, records=data)
     return out, []
@@ -51,8 +53,42 @@ def _generate_index_matching_statement(
     return base_txt.format(matching_set.name, domains_str, flattened.name)
 
 
+def _propagate_bounds(x, out):
+    """Propagate bounds from the input to the output variable"""
+
+    m = x.container
+
+    # set domain for variable
+    x_domain = x.domain
+    x_domain = utils._next_domains(x_domain, [])
+
+    bounds_set = m.addSet(records=["lb", "ub"])
+    bounds = m.addParameter(domain=[bounds_set, *x_domain])
+
+    # capture original bounds
+    bounds[("lb",) + tuple(x_domain)] = x.lo[x_domain]
+    bounds[("ub",) + tuple(x_domain)] = x.up[x_domain]
+
+    # reshape bounds based on the output variable's shape
+    # when bounds.records is None, it means the bounds are zeros
+    nb_data = (
+        None
+        if bounds.records is None
+        else bounds.toDense().reshape((2,) + out.shape)
+    )
+
+    # set new domain for bounds
+    nb_dom = [bounds_set, *out.domain]
+
+    new_bounds = m.addParameter(domain=nb_dom, records=nb_data)
+
+    # assign new bounds to the output variable
+    out.lo[...] = new_bounds[("lb",) + tuple(out.domain)]
+    out.up[...] = new_bounds[("ub",) + tuple(out.domain)]
+
+
 def _flatten_dims_var(
-    x: gp.Variable, dims: list[int]
+    x: gp.Variable, dims: list[int], propagate_bounds: bool = True
 ) -> tuple[gp.Variable, list[gp.Equation]]:
     m = x.container
     new_domain, flattened = _get_new_domain(x, dims)
@@ -60,6 +96,10 @@ def _flatten_dims_var(
     out = m.addVariable(
         domain=new_domain
     )  # outputs domain nearly matches the input domain
+
+    if propagate_bounds and x.records is not None:
+        _propagate_bounds(x, out)
+
     # match the flattened set to correct dims
     forwarded_domain = utils._next_domains([flattened, *x.domain], [])
     doms_to_flatten = [forwarded_domain[d + 1] for d in dims]
@@ -85,10 +125,14 @@ def _flatten_dims_var(
 
 
 def flatten_dims(
-    x: gp.Variable | gp.Parameter, dims: list[int]
+    x: gp.Variable | gp.Parameter,
+    dims: list[int],
+    propagate_bounds: bool = True,
 ) -> tuple[gp.Parameter | gp.Variable, list[gp.Equation]]:
     """
     Flatten domains indicated by `dims` into a single domain.
+    If `propagate_bounds` is True, and `x` is of type variable,
+    the bounds of the input variable are propagated to the output.
 
     Parameters
     ----------
@@ -97,6 +141,8 @@ def flatten_dims(
     dims: list[int]
         List of integers indicating indices of the domains
         to be flattened. Must be consecutive indices.
+    propagate_bounds: bool, optional
+        Propagate bounds from the input to the output variable. Default is True.
 
     Examples
     --------
@@ -113,6 +159,9 @@ def flatten_dims(
     """
     if not isinstance(x, (gp.Parameter, gp.Variable)):
         raise ValidationError("Expected a parameter or a variable input")
+
+    if not isinstance(propagate_bounds, bool):
+        raise ValidationError("Expected a boolean for propagate_bounds")
 
     if len(dims) < 2:
         raise ValidationError("Expected at least 2 items in the dim array")
@@ -141,4 +190,4 @@ def flatten_dims(
     if isinstance(x, gp.Parameter):
         return _flatten_dims_par(x, dims)
 
-    return _flatten_dims_var(x, dims)
+    return _flatten_dims_var(x, dims, propagate_bounds)