+++ tests/integration/test_model_instance.py
@@ -3,9 +3,14 @@ from __future__ import annotations
 import glob
 import math
 import os
+import platform
+import subprocess
+import sys
 
+import gamspy_base
 import pandas as pd
 import pytest
+from gams import GamsExceptionExecution
 
 from gamspy import (
     Container,
@@ -18,6 +23,7 @@ from gamspy import (
     Product,
     Sense,
     Set,
+    SolveStatus,
     Sum,
     Variable,
 )
@@ -25,11 +31,20 @@ from gamspy.exceptions import ValidationError
 
 pytestmark = pytest.mark.integration
 
+try:
+    from dotenv import load_dotenv
+
+    load_dotenv(os.getcwd() + os.sep + ".env")
+except Exception:
+    pass
+
 
 @pytest.fixture
 def data():
+    ci_license_path = os.path.join(gamspy_base.directory, "ci_license.txt")
+
     # Arrange
-    m = Container()
+    m = Container(debugging_level="keep")
     canning_plants = ["seattle", "san-diego"]
     markets = ["new-york", "chicago", "topeka"]
     distances = [
@@ -43,11 +58,46 @@ def data():
     capacities = [["seattle", 350], ["san-diego", 600]]
     demands = [["new-york", 325], ["chicago", 300], ["topeka", 275]]
 
+    process = subprocess.run(
+        [
+            sys.executable,
+            "-m",
+            "gamspy",
+            "install",
+            "license",
+            os.environ["MODEL_INSTANCE_LICENSE"],
+            "-o",
+            ci_license_path,
+        ],
+        check=True,
+        capture_output=True,
+        text=True,
+    )
+
+    assert process.returncode == 0, process.stderr
+
     # Act and assert
     yield m, canning_plants, markets, capacities, demands, distances
 
-    # Cleanup
     m.close()
+
+    process = subprocess.run(
+        [
+            sys.executable,
+            "-m",
+            "gamspy",
+            "install",
+            "license",
+            os.environ["LOCAL_LICENSE"],
+            "-o",
+            ci_license_path,
+        ],
+        check=True,
+        capture_output=True,
+        text=True,
+    )
+
+    assert process.returncode == 0, process.stderr
 
     files = glob.glob("_*")
     for file in files:
@@ -60,6 +110,10 @@ def data():
         os.remove("gams.gms")
 
 
+@pytest.mark.skipif(
+    platform.system() == "Darwin",
+    reason="Darwin runners are not dockerized yet.",
+)
 def test_parameter_change(data):
     m, canning_plants, markets, capacities, demands, distances = data
     i = Set(m, name="i", records=canning_plants)
@@ -142,6 +196,10 @@ def test_parameter_change(data):
     assert not transport._is_frozen
 
 
+@pytest.mark.skipif(
+    platform.system() == "Darwin",
+    reason="Darwin runners are not dockerized yet.",
+)
 def test_variable_change(data):
     m, canning_plants, markets, capacities, demands, distances = data
     i = Set(m, name="i", records=canning_plants)
@@ -185,6 +243,10 @@ def test_variable_change(data):
     transport.unfreeze()
 
 
+@pytest.mark.skipif(
+    platform.system() == "Darwin",
+    reason="Darwin runners are not dockerized yet.",
+)
 def test_fx(data):
     m, *_ = data
     INCOME0 = Parameter(
@@ -230,6 +292,10 @@ def test_fx(data):
     mm.unfreeze()
 
 
+@pytest.mark.skipif(
+    platform.system() == "Darwin",
+    reason="Darwin runners are not dockerized yet.",
+)
 def test_validations(data):
     m, canning_plants, markets, capacities, demands, distances = data
     i = Set(m, name="i", records=canning_plants)
@@ -282,10 +348,23 @@ def test_validations(data):
     assert os.path.exists("gams.gms")
 
     # Test solver options
-    transport.solve(solver="conopt", solver_options={"rtmaxv": "1.e12"})
-    assert os.path.exists(os.path.join(m.working_directory, "conopt.opt"))
+    with open("_out.txt", "w") as file:
+        transport.solve(
+            solver="conopt", output=file, solver_options={"rtmaxv": "1.e12"}
+        )
+
+    with open("_out.txt") as file:
+        assert ">>  rtmaxv 1.e12" in file.read()
+
+    options_path = os.path.join(m.working_directory, "conopt.opt")
+    assert os.path.exists(options_path)
+    os.remove(options_path)
 
 
+@pytest.mark.skipif(
+    platform.system() == "Darwin",
+    reason="Darwin runners are not dockerized yet.",
+)
 def test_modifiable_in_condition(data):
     m, *_ = data
     td_data = pd.DataFrame(
@@ -511,6 +590,10 @@ def test_modifiable_in_condition(data):
         war.freeze(modifiables=[x.l])
 
 
+@pytest.mark.skipif(
+    platform.system() == "Darwin",
+    reason="Darwin runners are not dockerized yet.",
+)
 def test_modifiable_with_domain(data):
     m, *_ = data
     import gamspy as gp
@@ -547,4 +630,75 @@ def test_modifiable_with_domain(data):
     mymodel.solve()
     assert math.isclose(
         mymodel.objective_value, 32.36124699832342, rel_tol=1e-6
+    )
+
+
+@pytest.mark.skipif(
+    platform.system() == "Darwin",
+    reason="Darwin runners are not dockerized yet.",
+)
+def test_license():
+    ci_license_path = os.path.join(gamspy_base.directory, "ci_license.txt")
+    subprocess.run(
+        [
+            sys.executable,
+            "-m",
+            "gamspy",
+            "uninstall",
+            "license",
+        ],
+        check=True,
+        capture_output=True,
+        text=True,
+    )
+    m = Container()
+    i = Set(m, "i", records=range(5000))
+    p = Parameter(m, "p", domain=i)
+    p2 = Parameter(m, "p2", records=5)
+    p.generateRecords()
+    v1 = Variable(m, "v1", domain=i)
+    z = Variable(m, "z")
+    e1 = Equation(m, "e1", domain=i)
+
+    e1[i] = p2 * v1[i] * p[i] >= z
+    model = Model(
+        m, name="my_model", equations=[e1], sense=Sense.MIN, objective=z
+    )
+    with pytest.raises(GamsExceptionExecution):
+        model.freeze(modifiables=[p2])
+
+    subprocess.run(
+        [
+            sys.executable,
+            "-m",
+            "gamspy",
+            "install",
+            "license",
+            os.environ["MODEL_INSTANCE_LICENSE"],
+            "-o",
+            ci_license_path,
+        ],
+        check=True,
+        capture_output=True,
+        text=True,
+    )
+
+    model.freeze(modifiables=[p2])
+    model.solve()
+    assert model.solve_status == SolveStatus.NormalCompletion
+
+    subprocess.run(
+        [
+            sys.executable,
+            "-m",
+            "gamspy",
+            "install",
+            "license",
+            os.environ["LOCAL_LICENSE"],
+            "-o",
+            ci_license_path,
+        ],
+        check=True,
+        capture_output=True,
+        text=True,
     )